feat: Replace PostSuccessDialog with PostStatusDialog and unify status handling in WritePage
