+++ app/feed/new/page.tsx
@@ -5,7 +5,7 @@ import ProtectedRoute from "@/components/auth/ProtectedRoute";
 import Background from "@/components/common/Background";
 import ErrorMessage from "@/components/post/ErrorMessageDialog";
 import { PostForm } from "@/components/post/PostForm";
-import PostSuccessDialog from "@/components/post/PostSuccessDialog";
+import PostStatusDialog from "@/components/post/PostStatusDialog";
 import { editorStrings } from "@/constants/editor";
 import {
   PostErrorResponse,
@@ -23,14 +23,16 @@ import { useCallback, useState } from "react";
 
 export default function WritePage() {
   const router = useRouter();
-  const [isSubmitting, setIsSubmitting] = useState(false);
+  // unified status: idle | submitting | onchain | completed | error
+  const [status, setStatus] = useState<
+    "idle" | "submitting" | "onchain" | "completed" | "error"
+  >("idle");
   const [createdPost, setCreatedPost] = useState<
     PostResponse | PostErrorResponse | null
   >(null);
   const [errorMessage, setErrorMessage] = useState<string>("");
   const [isSuccessDialogOpen, setIsSuccessDialogOpen] =
     useState<boolean>(false);
-  const [isErrorDialogOpen, setIsErrorDialogOpen] = useState<boolean>(false);
 
   const { createPost, submitToBlockchain, processIpfs } = useKonthoKoshApi();
 
@@ -39,15 +41,15 @@ export default function WritePage() {
   const [onChainProcessResp, setOnChainProcessResp] =
     useState<BlockchainProcessResponse | null>(null);
   const [onChainStatusMessage, setOnChainStatusMessage] = useState<string>("");
-  const [isOnChainProcessing, setIsOnChainProcessing] =
-    useState<boolean>(false);
 
   /**
    * Handle form submission (Direct KonthoKosh API)
    */
   const handleSubmit = useCallback(
     async (data: CreatePostRequest) => {
-      setIsSubmitting(true);
+      // open dialog immediately and move to submitting status
+      setIsSuccessDialogOpen(true);
+      setStatus("submitting");
       setErrorMessage("");
       setCreatedPost(null);
 
@@ -58,10 +60,8 @@ export default function WritePage() {
 
         // Success when we receive an object with `id`
         if (newPost && "id" in newPost) {
-          setIsSuccessDialogOpen(true);
-
-          // Start on-chain flow: submit then process serially.
-          setIsOnChainProcessing(true);
+          // start on-chain flow
+          setStatus("onchain");
           setOnChainStatusMessage("Submitting to blockchain...");
           try {
             const submitResp = await submitToBlockchain(newPost.id);
@@ -71,33 +71,28 @@ export default function WritePage() {
             );
 
             try {
-              // Use the blockchain entry id returned by submit response
               const processResp = await processIpfs(newPost.id);
               setOnChainProcessResp(processResp);
               setOnChainStatusMessage("On-chain processing completed.");
+              setStatus("completed");
             } catch (procErr) {
               const friendly = handleKonthoKoshError(procErr);
               setOnChainStatusMessage(friendly);
+              setStatus("error");
             }
           } catch (submitErr) {
             const friendly = handleKonthoKoshError(submitErr);
             setOnChainStatusMessage(friendly);
-          } finally {
-            setIsOnChainProcessing(false);
+            setStatus("error");
           }
         } else {
-          setIsErrorDialogOpen(true);
-          setErrorMessage(
-            handleKonthoKoshError({
-              status: 409,
-            })
-          );
+          setErrorMessage(handleKonthoKoshError({ status: 409 }));
+          setStatus("error");
         }
       } catch (error) {
         const friendly = handleKonthoKoshError(error);
         setErrorMessage(friendly);
-      } finally {
-        setIsSubmitting(false);
+        setStatus("error");
       }
     },
     [createPost, submitToBlockchain, processIpfs]
@@ -120,6 +115,28 @@ export default function WritePage() {
     [createPost]
   );
 
+  // Reset all submission related states when dialog is closed
+  const handleSuccessDialogOpenChange = useCallback(
+    (value: React.SetStateAction<boolean>) => {
+      setIsSuccessDialogOpen((prev) => {
+        const newOpen =
+          typeof value === "function"
+            ? (value as (p: boolean) => boolean)(prev)
+            : value;
+        if (!newOpen) {
+          setStatus("idle");
+          setCreatedPost(null);
+          setErrorMessage("");
+          setOnChainSubmitResp(null);
+          setOnChainProcessResp(null);
+          setOnChainStatusMessage("");
+        }
+        return newOpen;
+      });
+    },
+    []
+  );
+
   return (
     <ProtectedRoute>
       <Background>
@@ -131,26 +148,18 @@ export default function WritePage() {
               <PostForm
                 onSubmit={handleSubmit}
                 onSaveDraft={handleSaveDraft}
-                isSubmitting={isSubmitting}
+                isSubmitting={status === "submitting" || status === "onchain"}
               />
             </div>
 
-            <PostSuccessDialog
+            <PostStatusDialog
               isOpen={isSuccessDialogOpen}
-              onOpenChange={setIsSuccessDialogOpen}
+              onOpenChange={handleSuccessDialogOpenChange}
               createdPost={createdPost as PostResponse}
               onChainSubmit={onChainSubmitResp}
               onChainProcess={onChainProcessResp}
-              onChainStatusMessage={onChainStatusMessage}
-              isOnChainProcessing={isOnChainProcessing}
-              isSubmitting={isSubmitting}
-            />
-
-            <ErrorMessage
-              isOpen={isErrorDialogOpen}
-              onOpenChange={setIsErrorDialogOpen}
-              message={errorMessage}
-              postError={createdPost as PostErrorResponse | null}
+              status={status}
+              statusMessage={onChainStatusMessage || errorMessage}
             />
           </div>
         </main>
